Support /me as a valid route for current user
