+++ src/views/me/index.js
@@ -1,0 +1,6 @@
+// @flow
+import User from '../user';
+import UserSettings from '../userSettings';
+
+export default User;
+export const MeSettings = UserSettings;
