+++ src/views/navbar/components/profileDropdown.js
@@ -46,7 +46,7 @@ export const ProfileDropdown = (props: ProfileProps) => {
     <UserProfileDropdown className={'dropdown'}>
       <UserProfileDropdownList>
         {props.user.username && (
-          <Link rel="nofollow" to={`/users/${props.user.username}/settings`}>
+          <Link rel="nofollow" to={`/me/settings`}>
             <UserProfileDropdownListItem>
               My Settings
             </UserProfileDropdownListItem>
